third_party/blink/web_tests/editing/deleting/delete-and-cleanup-expected.txt

This test checks that deletion does not leave unnecessary nested divs.

PASS confirmedMarkup is '<br>'
PASS confirmedMarkup is '<br>'
PASS confirmedMarkup is '<div id="mydiv"><br></div>'
PASS confirmedMarkup is '<br>'
PASS confirmedMarkup is '<b><br></b>'
PASS confirmedMarkup is '<b><div style="border: solid red"><br></div></b>'
PASS confirmedMarkup is '<br>'
PASS confirmedMarkup is '<br>'
PASS confirmedMarkup is '<p><br></p>'
PASS successfullyParsed is true

TEST COMPLETE
